requirements.txt
langgraph>=0.2.0
langchain-groq>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
duckduckgo-search>=8.0.0
pytest>=8.0.0
